[refactor(folder)] molecule -> atoms 변경
