+++ src/components/atoms/Form/FormInput/index.stories.tsx
