+++ src/components/atoms/Pagination/style.scss
