+++ src/components/atoms/SideNavigation/SideNavigationItem/style.scss
