+++ src/components/atoms/SideNavigation/style.scss
