+++ src/index.scss
@@ -49,14 +49,6 @@ button {
     cursor: pointer;
 }
 
-select {
-    -webkit-appearance: inherit;
-    background: url("./components/molecules/FormSelect/downarrow.png") no-repeat
-        95% 50%;
-    background-size: 1rem;
-    border-radius: 0;
-}
-
 ul {
     list-style: none;
 }
